Visual Studio 2019 compatibility with minimal changes, should compile fully with no warnings or errors. Modify project properties to use MSVC 142. Projects are still compatible with all VS versions down to VS 2010. Miscellaneous compatibility fixes codebase-wide. Add Win32 manifest file for better compatibility. Set minimum Windows version to WinVista. Add #pragma warnings to ignore unimportant warnings introduced by VS versions newer than 2010. Add .gitignore. Add .gitattributes.

+++ neo/idlib/math/Quat.h
@@ -285,7 +285,7 @@ ID_INLINE idQuat& idQuat::Normalize() {
 
 ID_INLINE float idQuat::CalcW() const {
 	// take the absolute value because floating point rounding may cause the dot of x,y,z to be larger than 1
-	return sqrt( fabs( 1.0f - ( x * x + y * y + z * z ) ) );
+	return sqrtf( fabsf( 1.0f - ( x * x + y * y + z * z ) ) );
 }
 
 ID_INLINE int idQuat::GetDimension() const {
@@ -407,7 +407,7 @@ ID_INLINE int idCQuat::GetDimension() const {
 
 ID_INLINE idQuat idCQuat::ToQuat() const {
 	// take the absolute value because floating point rounding may cause the dot of x,y,z to be larger than 1
-	return idQuat( x, y, z, sqrt( fabs( 1.0f - ( x * x + y * y + z * z ) ) ) );
+	return idQuat( x, y, z, sqrtf( fabsf( 1.0f - ( x * x + y * y + z * z ) ) ) );
 }
 
 ID_INLINE const float *idCQuat::ToFloatPtr() const {
